LDEV-755 Break routines out of calculate_atmospheric_parameters.c
The calculate_atmospheric_parameters.c file is over 2000 lines, consisting
of several sets of functions that should probably be in separate files
based on functionality.

Also, reduce the number of memory allocation and freeing for work arrays
by using structures that can be allocated once and freed when processing
is complete.